update to the new Languages api
