+++ REQUIRE
@@ -2,5 +2,4 @@ julia 0.6
 BinaryProvider
 Languages 0.2.0
 DataFrames
-BinDeps
-Compat
+WordTokenizers
