+++ src/ngramizer.jl
@@ -4,8 +4,8 @@
 #
 ##############################################################################
 
-function ngramize{S <: Language, T <: AbstractString}(::Type{S}, words::Vector{T}, n::Int)
-    (n == 1) && return onegramize(S, words)
+function ngramize{S <: Language, T <: AbstractString}(lang::S, words::Vector{T}, n::Int)
+    (n == 1) && return onegramize(lang, words)
 
     n_words = length(words)
 
@@ -21,7 +21,7 @@ function ngramize{S <: Language, T <: AbstractString}(::Type{S}, words::Vector{T
     return tokens
 end
 
-function onegramize{S <: Language, T <: AbstractString}(::Type{S}, words::Vector{T})
+function onegramize{S <: Language, T <: AbstractString}(lang::S, words::Vector{T})
     n_words = length(words)
     tokens = Dict{T, Int}()
