+++ src/preprocessing.jl
@@ -158,46 +158,6 @@ function remove_words!{T <: AbstractString}(entity::(Union{AbstractDocument,Corp
     prepare!(entity, strip_patterns, skip_words = skipwords)
 end
 
-function remove_whitespace!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_whitespace! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_whitespace)
-end
-function remove_punctuation!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_punctuation! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_punctuation)
-end
-function remove_nonletters!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_nonletters! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_non_letters)
-end
-function remove_numbers!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_numbers! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_numbers)
-end
-function remove_articles!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_articles! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_articles)
-end
-function remove_indefinite_articles!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_indefinite_articles! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_indefinite_articles)
-end
-function remove_definite_articles!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_definite_articles! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_definite_articles)
-end
-function remove_prepositions!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_prepositions! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_prepositions)
-end
-function remove_pronouns!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_pronouns! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_pronouns)
-end
-function remove_stop_words!(entity::(Union{AbstractDocument,Corpus}))
-    Base.warn_once("remove_stop_words! is deprecated, Use prepare! instead.")
-    prepare!(entity, strip_stopwords)
-end
 
 
 ##############################################################################
@@ -305,7 +265,7 @@ function remove_patterns(s::AbstractString, rex::Regex)
     String(take!(iob))
 end
 
-function remove_patterns{T <: Compat.String}(s::SubString{T}, rex::Regex)
+function remove_patterns{T <: String}(s::SubString{T}, rex::Regex)
     iob = IOBuffer()
     ioffset = s.offset
     data = s.string.data
