+++ test/ngramizer.jl
@@ -5,15 +5,15 @@ module TestNGramizer
     using Compat
 
     sample_text = "this is some sample text"
-    tkns = TextAnalysis.tokenize(Languages.EnglishLanguage, sample_text)
-    ngs = TextAnalysis.ngramize(TextAnalysis.EnglishLanguage, tkns, 1)
-    @assert isequal(ngs, Dict{Compat.UTF8String,Int}("some" => 1,
+    tkns = TextAnalysis.tokenize(Languages.English(), sample_text)
+    ngs = TextAnalysis.ngramize(Languages.English(), tkns, 1)
+    @assert isequal(ngs, Dict{String,Int}("some" => 1,
     	                                     "this" => 1,
     	                                     "is" => 1,
     	                                     "sample" => 1,
     	                                     "text" => 1))
-    ngs = TextAnalysis.ngramize(TextAnalysis.EnglishLanguage, tkns, 2)
-    @assert isequal(ngs, Dict{Compat.UTF8String,Int}("some" => 1,
+    ngs = TextAnalysis.ngramize(Languages.English(), tkns, 2)
+    @assert isequal(ngs, Dict{String,Int}("some" => 1,
                                              "this is" => 1,
                                              "some sample" => 1,
                                              "is some" => 1,
