+++ test/tokenizer.jl
@@ -2,17 +2,17 @@ module TestTokenizer
     using Base.Test
     using Languages
     using TextAnalysis
-    using Compat
+
 
     sample_text = "this is some sample text"
 
     tkns = TextAnalysis.tokenize(
-        Languages.EnglishLanguage,
+        Languages.English(),
         sample_text
     )
 
     @assert isequal(
         tkns,
-        Compat.UTF8String["this", "is", "some", "sample", "text"]
+        String["this", "is", "some", "sample", "text"]
     )
 end
